Routing (#6)
* Update file paths, modify SQL queries, and remove login/register pages

- Change file paths to include ../ for more accurate directory referencing
- Modify SQL queries to insert cause areas for opportunities and update opportunity_cause_areas table
- Remove login and register pages from the project

* Backend

* "Added session_start() and user_id check in core.php; Changed cause_area to cause_area[] in post_opportunity.php; Updated post_opportunity_process.php to handle user_id and cause_area[]"

* Added dynamic rendering of volunteer opportunities, fetching data from the database

This commit adds dynamic rendering of volunteer opportunities by fetching data from the database. It includes a new PHP script to connect to the database and query for opportunities with status 1. The opportunities are then displayed in the HTML using a while loop to iterate through the query results. If no opportunities are available, a message is displayed. The database connection is closed after use.